requirements.txt
anyio==3.6.2
autopep8==2.0.1
breadability==0.1.20
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
fastapi==0.88.0
ffmpeg-python==0.2.0
filelock==3.8.2
future==0.18.2
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
lxml==4.9.2
more-itertools==9.0.0
nltk==3.8
numpy==1.24.0
packaging==22.0
pycodestyle==2.10.0
pycountry==22.3.5
pydantic==1.10.2
python-dotenv==0.21.0
pytube==12.1.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
sumy==0.11.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
whisperx @ git+https://github.com/m-bain/whisperx.git@c912f96ed35f59a602075c817510403f3f1fa9e7
